test(ios): MIM-10 拆分中断回归测试
